add trackClick
