+++ packages/web/src/track-click.ts
@@ -1,0 +1,5 @@
+import { TrackClickInput } from './types';
+
+export const trackClick = (event: TrackClickInput) => {
+  console.log('trackClick', event);
+};
